validationtools/pyproject.toml
[project]
name = "valiationtools"
version = "0.1"
dependencies = ["pyserial", "protobuf"]